==> 2.Data/01_Code_ProvMun_11_21.ipynb <==
## Códigos CCAA provincias y municipios 2011 y 2021
En  este notebook procesaremos dos .xlsx con la relación de provincias y municipios con sus códigos del INE.  
Uno de ellos es relativo al año 2011 y el otro al año 2021.  
Obtendremos dos DataFrame con la siguiente información:
- Nombre CCAA
- Codigo CCAA
- Nombre Provincia
- Codigo Provincia
- Nombre Municipio
- Codigo Municipio: se compone de cinco dígitos, los dos primeros corresponden al código de la provincia y los tres restantes al del municipio dentro de ésta

import pandas as pd

Tabla que relaciona códigos de CCAA, provincias y municipios en 2011:

ProvMun = pd.read_excel('01_Input_ProvMun_11.xlsx', dtype = {'CODAUTO' : str, 'CPRO' : str, 'CMUN' : str})

ProvMun['Codigo Municipio'] = ProvMun['CPRO'].map(str) + ProvMun['CMUN'].map(str)

ProvMun = ProvMun.drop(['CMUN','DC'], axis = 1)

ProvMun.rename({'NOMBREAUTO' : 'Nombre CCAA',
                'CODAUTO' : 'Codigo CCAA',
                'CPRO': 'Codigo Provincia', 
                'NOMBREPRO':'Nombre Provincia',
                'NOMBREMUN': 'Nombre Municipio' }, axis = 1, inplace = True)

ProvMun = ProvMun.reindex(columns = ['Nombre CCAA', 
                                   'Codigo CCAA', 
                                   'Nombre Provincia',
                                   'Codigo Provincia',
                                   'Nombre Municipio',
                                   'Codigo Municipio'])

ProvMun.head()

ProvMun.to_csv('01_Output_ProvMun_11.csv', header = True, index = False)

Tabla que relaciona códigos de CCAA provincias y municipios en 2021:

ProvMun = pd.read_excel('01_Input_ProvMun_21.xlsx', dtype = {'CODAUTO' : str, 'CPRO' : str, 'CMUN' : str}) 

ProvMun['Codigo Municipio'] = ProvMun['CPRO'].map(str) + ProvMun['CMUN'].map(str)

ProvMun = ProvMun.drop(['CMUN','DC'], axis = 1)

ProvMun.rename({'NOMBREAUTO' : 'Nombre CCAA',
                'CODAUTO' : 'Codigo CCAA',
                'CPRO': 'Codigo Provincia', 
                'NOMBREPRO':'Nombre Provincia',
                'NOMBREMUN': 'Nombre Municipio' }, axis=1, inplace=True)

ProvMun = ProvMun.reindex(columns = ['Nombre CCAA', 
                                   'Codigo CCAA', 
                                    'Nombre Provincia',
                                    'Codigo Provincia',
                                    'Nombre Municipio',
                                    'Codigo Municipio'])

ProvMun.head()

ProvMun.to_csv('01_Output_ProvMun_21.csv', header = True, index = False)
