==> 2.Data/02_Code_Poblacion_11_21.ipynb <==
## Población 2011 y 2021

En este notebook procesaremos 52 JSON, uno por cada provincia y ciudad autónoma.  
Cada uno de ellos contiene el total de población por municipios así como la segmentación por sexo. 

import pandas as pd
import requests

Creamos una lista con el número de tabla del INE correspondiente a cada provincia y ciudad autónoma:

CodTablas = [2854, 2855, 2856, 2857, 2858, 2859, 2860, 2861, 2862, 2863, 2864, 2865, 2866, 2868, 2869, 2870,2871,
2872, 2873,  2874, 2875, 2876, 2877, 2878, 2879, 2880, 2881, 2882, 2883, 2884, 2885, 2886, 2888, 2889, 2890, 2891, 
2892, 2893, 2894, 2895, 2896, 2899, 2900, 2901,2902, 2903, 2904, 2905, 2906, 2907,  2908, 2909]

def Ine(ine_code):
    
    Path = 'https://servicios.ine.es/wstempus/js/es/DATOS_TABLA/{cod_serie}?tip=AM&'
    Path = Path.format(cod_serie=ine_code)
    json_request = requests.get(Path).json()
    
    return json_request

### Población 2011
 
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Poblacion Total
- Poblacion Hombres
- Poblacion Mujeres

CodHomList = list()
PobHomList = list()
PobMujList = list()
CodMujList = list()
PobTotList = list()
CodTolList = list()

Nos quedamos solo con el año 2011:

for codigo in CodTablas:
    datos = Ine(codigo)
    for dato in datos: 
        for i in range(0,len(dato['Data'])):
            if  dato['Data'][i]['Anyo'] == 2011 and 'Hombres' in dato['Nombre']:
                PobHomList.append(dato['Data'][i]['Valor'])
                CodHomList.append(dato['MetaData'][0]['Codigo'])
            elif dato['Data'][i]['Anyo'] == 2011 and 'Mujeres' in dato['Nombre']:
                PobMujList.append(dato['Data'][i]['Valor'])
                CodMujList.append(dato['MetaData'][0]['Codigo'])
            elif dato['Data'][i]['Anyo'] == 2011 and ('Hombres'and'Mujeres') not in dato['Nombre']:
                PobTotList.append(dato['Data'][i]['Valor'])
                CodTolList.append(dato['MetaData'][0]['Codigo'])            

PoblacionHombres = pd.DataFrame({
        'Codigo Municipio' : CodHomList,
        'Poblacion Hombre': PobHomList
    })

PoblacionMujeres = pd.DataFrame({
        'Codigo Municipio' : CodMujList,
        'Poblacion Mujer': PobMujList
    })

PoblacionTotal = pd.DataFrame({
        'Codigo Municipio' : CodTolList,
        'Poblacion Total': PobTotList
    })

Poblacion11 = pd.concat([PoblacionTotal, PoblacionHombres, PoblacionMujeres], axis = 1)
Poblacion11 = Poblacion11.loc[:, ~Poblacion11.columns.duplicated()]

Poblacion11.head()

Poblacion11.to_csv('02_Output_Poblacion_11.csv', header = True, index = False)

### Población 2021
 
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Poblacion Total
- Poblacion Hombres
- Poblacion Mujeres

CodHomList = list()
PobHomList = list()
PobMujList = list()
CodMujList = list()
PobTotList = list()
CodTolList = list()

Nos quedamos solo con el año 2021:

for codigo in CodTablas:
    datos = Ine(codigo)
    for dato in datos: 
        for i in range(0,len(dato['Data'])):
            if  dato['Data'][i]['Anyo'] == 2021 and 'Hombres' in dato['Nombre']:
                PobHomList.append(dato['Data'][i]['Valor'])
                CodHomList.append(dato['MetaData'][0]['Codigo'])
            elif dato['Data'][i]['Anyo'] == 2021 and 'Mujeres' in dato['Nombre']:
                PobMujList.append(dato['Data'][i]['Valor'])
                CodMujList.append(dato['MetaData'][0]['Codigo'])
            elif dato['Data'][i]['Anyo'] == 2021 and ('Hombres'and'Mujeres') not in dato['Nombre']:
                PobTotList.append(dato['Data'][i]['Valor'])
                CodTolList.append(dato['MetaData'][0]['Codigo'])            

PoblacionHombres = pd.DataFrame({
        'Codigo Municipio' : CodHomList,
        'Poblacion Hombre': PobHomList
    })

PoblacionMujeres = pd.DataFrame({
        'Codigo Municipio' : CodMujList,
        'Poblacion Mujer': PobMujList
    })

PoblacionTotal = pd.DataFrame({
        'Codigo Municipio' : CodTolList,
        'Poblacion Total': PobTotList
    })

Poblacion21 = pd.concat([PoblacionTotal, PoblacionHombres, PoblacionMujeres], axis = 1)
Poblacion21 = Poblacion21.loc[:, ~Poblacion21.columns.duplicated()]

Poblacion21.head()

Poblacion21.to_csv('02_Output_Poblacion_21.csv', header = True, index = False)
