==> 2.Data/03_Code_Densidad_poblacion_11_21.ipynb <==
## Densidad población 2011 y 2021

En este notebook procesaremos un .xlsx con la superficie de cada uno de los municipios junto con los csv files que contienen los datos de población.  
Calculamos la densidad desagregada por municipios y ciudades autónomas para 2011 y 2021.

import pandas as pd

DensidadPoblacion = pd.read_excel('03_Input_Densidad_Poblacion_11_21.xlsx',
                                  dtype = {'ID_REL' : str}, 
                                  usecols = ('ID_REL','SUPERFICIE'))

DensidadPoblacion.head()

DensidadPoblacion.rename({'ID_REL': 'Codigo Municipio'}, axis = 1, inplace = True)

### Densidad población 2011

Junto con la tabla 02_Output_Poblacion_11.csv que contiene la población de 2011, obtendremos un DataFrame 
con la siguiente información:
- Codigo Municipio
- Superficie km2
- Densidad Población

Poblacion11 = pd.read_csv('02_Output_Poblacion_11.csv',
                          dtype = {'Codigo Municipio' : str},
                          usecols = ( 'Codigo Municipio', 'Poblacion Total'))

Poblacion11.head()

DensidadPoblacion11 = DensidadPoblacion.merge(Poblacion11, on ='Codigo Municipio')

DensidadPoblacion11.head()

La superficie está en hectáreas, la pasamos a km2 y calculamos la densidad:

DensidadPoblacion11.dtypes

DensidadPoblacion11['Superficie Km2'] = DensidadPoblacion11['SUPERFICIE'].map(lambda x: x/100).round(2)

DensidadPoblacion11['Densidad Poblacion'] = (DensidadPoblacion11['Poblacion Total']\
                                             / DensidadPoblacion11['Superficie Km2']).round(2)

DensidadPoblacion11 = DensidadPoblacion11.drop(['SUPERFICIE', 'Poblacion Total'], axis = 1)

DensidadPoblacion11.head()

DensidadPoblacion11.to_csv('03_Output_Densidad_Poblacion_11.csv', header = True, index = False)

### Densidad población 2021

Junto con la tabla 02_Output_Poblacion_21.csv que contiene la población de 2021, obtendremos un DataFrame 
con la siguiente información:
- Codigo Municipio
- Superficie km2
- Densidad Población

Poblacion21 = pd.read_csv('02_Output_Poblacion_21.csv', 
                          dtype = {'Codigo Municipio' : str},
                          usecols = ( 'Codigo Municipio', 'Poblacion Total'))
Poblacion21.head()

DensidadPoblacion21 = DensidadPoblacion.merge(Poblacion21, on ='Codigo Municipio')

DensidadPoblacion21.head()

La superficie está en hectareas, la pasamos a km2 y calculamos la densidad:

DensidadPoblacion21['Superficie Km2'] = DensidadPoblacion21['SUPERFICIE'].map(lambda x: x/100).round(2)

DensidadPoblacion21['Densidad Poblacion'] = (DensidadPoblacion21['Poblacion Total']\
                                             / DensidadPoblacion21['Superficie Km2']).round(2)

DensidadPoblacion21 = DensidadPoblacion21.drop(['SUPERFICIE', 'Poblacion Total'],axis = 1)

DensidadPoblacion21.head()

DensidadPoblacion21.to_csv('03_Output_Densidad_Poblacion_21.csv', header = True, index = False)
