==> 2.Data/04_Code_Altitud_Municipios_11.ipynb <==
## Altitud municipios

En este notebook procesaremos el archivo 03_Input_Densidad_Poblacion_11_21.xlsx para extraer la variable altitud de los municipios.  
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Altitud Municipios

import pandas as pd

AltitudMunicipios = pd.read_excel('03_Input_Densidad_Poblacion_11_21.xlsx', 
                                  dtype = {'ID_REL' : str}, 
                                  usecols = ('ID_REL','ALTITUD'))

AltitudMunicipios.head()

AltitudMunicipios.rename({'ID_REL': 'Codigo Municipio',
                           'ALTITUD' : 'Altitud Municipios'}, axis = 1, inplace = True)

AltitudMunicipios.to_csv('04_Output_Altitud_Municipios_11.csv', header = True, index = False)
