==> 2.Data/05_Code_Nacimientos_Matrimonios_Defunciones_11.ipynb <==
## Nacimientos, matrimonios y defunciones 2011

En este notebook procesaremos 52 JSON, uno por cada provincia y ciudad autónoma.  
Cada uno de ellos contiene el total de nacimientos, matrimonios y defunciones desagregado por municipios en 2011.  
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio 
- Nacimientos
- Matrimonios
- Defunciones

import pandas as pd
import requests

Códigos de cada una de las tablas de las provincias y ciudades autónomas del INE correspondientes a los datos de 2011:

CodTablas = list(range(22808, 22859+1))

def Ine(ine_code):
    
    Path = 'https://servicios.ine.es/wstempus/js/es/DATOS_TABLA/{cod_serie}?tip=AM&'
    Path = Path.format(cod_serie=ine_code)
    json_request = requests.get(Path).json()
    
    return json_request

PobNacList = list()
NacList = list()
PobMatList = list()
MatList = list()
PobDefList = list()
DefList = list()

for codigo in CodTablas:
    datos = Ine(codigo)
    for dato in datos: 
        for i in range(0,len(dato['Data'])):
            if 'nacidos vivos por residencia materna' in dato['Nombre']:
                PobNacList.append(dato['Nombre'])
                NacList.append(dato['Data'][i]['Valor'])
            elif 'matrimonios por el lugar en que han fijado residencia' in dato['Nombre']:
                PobMatList.append(dato['Nombre'])
                MatList.append(dato['Data'][i]['Valor'])
            elif 'fallecidos por el lugar de residencia' in dato['Nombre']:
                PobDefList.append(dato['Nombre'])
                DefList.append(dato['Data'][i]['Valor']) 

Nacimientos = pd.DataFrame({
'Codigo Municipio' : PobNacList,
'Nacimientos': NacList
})

Matrimonios = pd.DataFrame({
'Codigo Municipio' : PobMatList,
'Matrimonios': MatList
})

Defunciones = pd.DataFrame({
'Codigo Municipio' : PobDefList,
'Defunciones': DefList
}) 


Naci_Matri_Defun11 = pd.concat([Nacimientos, Matrimonios, Defunciones], axis = 1)
Naci_Matri_Defun11 = Naci_Matri_Defun11.loc[:, ~Naci_Matri_Defun11.columns.duplicated()]

Naci_Matri_Defun11.head()

Cargamos la tabla 01_Output_ProvMun_11.csv para crear una columna con la variable Codigo Municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str, 
                                                         'Codigo Provincia' : str, 
                                                         'Codigo Municipio' : str}) 
ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

Naci_Matri_Defun11['Codigo Municipio'] = Naci_Matri_Defun11['Codigo Municipio'].map(lambda x: check_list(x))

Naci_Matri_Defun11.head()

Naci_Matri_Defun11.to_csv('05_Output_Nacimientos_Matrimonios_Defunciones_11.csv', header = True, index = False)
