==> 2.Data/06_Code_Afiliados_SS_11.ipynb <==
## Afiliados Seguridad Social 2011
En este notebook procesaremos un .xlsx con el número de afiliados por municipios, dentro del régimen general, agrario, hogar, mar, autónomos, carbón y una última columna con el total.  
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Total Regimen SS

import pandas as pd

AfiliadosSS11 = pd.read_excel('06_Input_Afiliados_SS_11.xlsx',dtype = {'CODIGO' : str})

AfiliadosSS11.head()

Los valores que están entre 0 y 5 se indican en la tabla con <5. Sustituimos estos valores por un 2.  
Ademá creamos una nueva columna con la suma de todas las columnas segun régimen.

AfiliadosSS11 = AfiliadosSS11.applymap(lambda x: 2 if x == '<5' else x)

AfiliadosSS11.dtypes

AfiliadosSS11['Total Regimen SS'] = AfiliadosSS11['GENERAL'] + AfiliadosSS11['AGRARIO'] \
                                    + AfiliadosSS11['HOGAR'] + AfiliadosSS11['MAR'] \
                                    + AfiliadosSS11['AUTONOMOS'] + AfiliadosSS11['CARBON']

AfiliadosSS11.head()

AfiliadosSS11.drop(['PROVINCIA', 
                    'MUNICIPIO', 
                    'TRAB', 
                    'GENERAL', 
                    'AGRARIO', 
                    'HOGAR', 
                    'MAR', 
                    'AUTONOMOS', 
                    'CARBON', 
                    'TOTAL'], axis = 1, inplace = True)

AfiliadosSS11.rename({'CODIGO' : 'Codigo Municipio'}, axis = 1, inplace = True)
AfiliadosSS11.head()

AfiliadosSS11.to_csv('06_Output_Afiliados_SS_11.csv', header = True, index = False)
