==> 2.Data/07_Code_Contratos_Laborales_11.ipynb <==
## Contratos laborales 2011

En este notebook procesaremos un .xls con el número de contratos laborales por municipios y por sexo.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Total Contratos
- Contratos Hombres
- Contrato Mujeres


import pandas as pd

ContratosLaborales11 = pd.read_excel('07_Input_Contratos_Laborales_11.xlsx', dtype = {'Codigo Municipio' : str})

ContratosLaborales11.head()

Calculamos el total de contratos de hombre y mujeres

ContratosLaborales11['Contratos Hombres'] = ContratosLaborales11['Contratos iniciales indefinidos hombres'] +\
                                            ContratosLaborales11['Contratos iniciales temporales hombres'] +\
                                            ContratosLaborales11['Contratos convertidos en indefinidos hombres']

ContratosLaborales11['Contratos Mujeres'] = ContratosLaborales11['Contratos iniciales indefinidos mujeres'] +\
                                            ContratosLaborales11['Contratos iniciales temporales mujeres'] +\
                                            ContratosLaborales11['Contratos convertidos en indefinidos mujeres']
ContratosLaborales11.head()

ContratosLaborales11.drop([' Municipio',
                           'Contratos iniciales indefinidos hombres', 
                           'Contratos iniciales temporales hombres', 
                           'Contratos convertidos en indefinidos hombres', 
                           'Contratos iniciales indefinidos mujeres', 
                           'Contratos iniciales temporales mujeres',
                           'Contratos convertidos en indefinidos mujeres',
                           'Contratos  Agricultura', 
                           'Contratos  Industria',
                           'Contratos Construcción',
                           'Contratos  Servicios'], axis = 1, inplace = True)

ContratosLaborales11.head()

ContratosLaborales11.to_csv('07_Output_Contratos_Laborales_11.csv', header = True, index = False)
