==> 2.Data/08_Code_Parados_11.ipynb <==
## Parados 2011
En este notebook procesaremos un .xlsx que contiene el paro registrado por municipios y sexo.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Total Paro Registrado
- Parados Hombres
- Parados Mujeres

import pandas as pd

Parados11 = pd.read_excel('08_Input_Parados_11.xlsx',dtype = {'Codigo Municipio' : str})

Parados11.head()

Calculamos el total de parados hombres y mujeres

Parados11['Parados Hombres'] = Parados11['Suma de Paro hombre edad < 25'] + Parados11['Suma de Paro hombre edad 25 -45 '] +\
                              Parados11['Suma de Paro hombre edad >=45']

Parados11['Parados Mujeres'] = Parados11['Suma de Paro mujer edad < 25'] + Parados11['Suma de Paro mujer edad 25 -45 '] +\
                              Parados11['Suma de Paro mujer edad >=45']

Parados11.head()

Parados11 = Parados11[['Codigo Municipio', 'Suma de total Paro Registrado', 'Parados Hombres', 'Parados Mujeres']]
Parados11.rename({'Suma de total Paro Registrado' : 'Total Paro Registrado'}, axis = 1, inplace = True)

Parados11.head()

Parados11.to_csv('08_Output_Parados_11.csv', header = True, index = False)
