==> 2.Data/09_Code_Numero_Establecimientos_11.ipynb <==
## Numero establecimientos 2011
En este notebook procesaremos un .xlsx con el número de establecimientos desagregado por municipios y por tipos de negocio: agricultura, industria, construcción y servicios. 

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº establecimientos Agricultura
- Nº establecimientos Industria
- Nº establecimientos Construccion
- Nº establecimientos Servicios


import pandas as pd
import numpy as np

NumeroEstablecimientos11 = pd.read_excel('09_Input_Numero_Establecimientos_11.xlsx', dtype = {'Codigo' : str})

NumeroEstablecimientos11.head()

NumeroEstablecimientos11.drop('Etiqueta', axis = 1, inplace = True)

NumeroEstablecimientos11.rename({'Código' : 'Codigo Municipio',
                                'Agricultura' : 'Nº Establecimientos Agricultura',
                                'Industria' : 'Nº Establecimientos Industria',
                                'Construcción ' : 'Nº Establecimientos Construccion', 
                                'Servicios' : 'Nº Establecimientos Servicios'}, axis = 1, inplace = True)

NumeroEstablecimientos11.head()

Algunas variables toman el valor: N/A - resultado no disponible  
Lo sustituimos por NaN:

NumeroEstablecimientos11.replace('N/A - resultado no disponible', np.nan, inplace = True)

NumeroEstablecimientos11.to_csv('09_Output_Numero_Establecimientos_11.csv', header = True, index = False)
