==> 2.Data/10_Code_Centros_Sanitarios_11.ipynb <==
## Numero centros sanitarios 2011
En este notebook procesaremos un .xlsx con dos hojas; una con el listado detallado de hospitales y otras con el listado detallado de centros de salud. Ambos desagregados por municipios. 

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº Centros Sanitarios


import pandas as pd

Procesamos el .xlsx de hospitales y de centros de salud del año 2011 y concatenamos:

Hospitales11 = pd.read_excel('10_Input_Centros_Sanitarios_11.xlsx', usecols = ['NOMBRE','CODMU'], 
                             sheet_name ='HOSPITALES', dtype = {'CODMU' : str})

Hospitales11.head()

CentrosSalud11 = pd.read_excel('10_Input_Centros_Sanitarios_11.xlsx', usecols = ['NOMBRE', 'CODMU'], 
                               sheet_name ='CENTROS', dtype = {'CODMU' : str})

CentrosSalud11.head()

CentrosSanitarios11 = pd.concat([Hospitales11, CentrosSalud11], ignore_index = True)

CentrosSanitarios11.head()

Contamos cuántos centros sanitarios hay por municipio agrupando por el código de municipio:

CentrosSanitarios11 = CentrosSanitarios11.groupby('CODMU')['NOMBRE'].count()
CentrosSanitarios11 = CentrosSanitarios11.reset_index()

CentrosSanitarios11.head()

CentrosSanitarios11.rename({'NOMBRE' : 'Nº Centros Sanitarios', 'CODMU' : 'Codigo Municipio'}, 
                           axis = 1, inplace = True)

CentrosSanitarios11.head()

CentrosSanitarios11.to_csv('10_Output_Centros_Sanitarios_11.csv', header = True, index = False)
