==> 2.Data/11_Code_Centros_Educativos_11.ipynb <==
## Numero de centros educativos 2011
En este notebook procesaremos un .xlsx  que contiene el listado detallado de centros educativos desagregados por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº Centros Educativos


import pandas as pd

CentrosEducativos11 = pd.read_excel('11_Input_Centros_Educativos_11.xlsx', 
                                    usecols = ['Denominación genérica', 'codigo ine'],
                                    dtype = {'codigo ine' : str})

CentrosEducativos11.head()

Contamos cuántos centros educativos hay por municipio agrupando por Codigo Ine:

CentrosEducativos11 = CentrosEducativos11.groupby('codigo ine')['Denominación genérica'].count()
CentrosEducativos11 = CentrosEducativos11.reset_index()

CentrosEducativos11.head()

CentrosEducativos11.rename({'codigo ine' : 'Codigo Municipio', 'Denominación genérica' : 'Nº Centros Educativos'},\
                           axis = 1, inplace = True )

CentrosEducativos11.head()

CentrosEducativos11.to_csv('11_Output_Centros_Educativos_11.csv', header = True, index = False)
