==> 2.Data/12_Code_Transacciones_Inmobiliarias_11.ipynb <==
## Número de transacciones inmobiliarias 2011
En este notebook procesaremos un .xlsx con el número de transacciones inmobiliarias de viviendas desagregado por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº Transacciones Inmobiliarias 


import pandas as pd
import numpy as np

Transacciones11 = pd.read_excel('12_Input_Transacciones_Inmobiliarias_11.xlsx', dtype = {'Código' : str})

Transacciones11.head()

Transacciones11.drop(['Etiqueta'], axis = 1, inplace = True)

Transacciones11.rename({'Código' : 'Codigo Municipio', 
                       'Número de transacciones anuales 2020' : 'Nº de transacciones'}, 
                       axis = 1, inplace = True)

Algunas variables toman el valor: N/A - resultado no disponible.  
Lo substituimos por NaN:

Transacciones11.replace('N/A - resultado no disponible', np.nan, inplace = True)

Transacciones11.to_csv('12_Output_Transacciones_Inmobiliarias_11.csv', header = True, index = False)
