==> 2.Data/13_Code_Establecimientos_Colectivos_11.ipynb <==
## Establecimientos colectivos 2011

En este notebook procesaremos un JSON que contiene el número de residencias de personas mayores, instituciones religiosas y militares desagregado por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº Residencias Mayores
- Nº Instituciones Religiosas y Militares

import pandas as pd
import requests

Path = 'https://servicios.ine.es/wstempus/js/es/DATOS_TABLA//t20/e244/colectivos/p05/l0/1mun00.px?tip=AM&p=2&'

json_request = requests.get(Path).json()

CodResiList = list()
NumResiList = list()
CodReliList = list()
NumReliList = list()

for dato in json_request: 
    for i in range(0,len(dato['Data'])):
        if 'Residencias de personas mayores' in dato['Nombre']:
            CodResiList.append(dato['MetaData'][0]['Codigo'])
            NumResiList.append(dato['Data'][i]['Valor'])
        elif 'Instituciones religiosas o militares' in dato['Nombre']:
            CodReliList.append(dato['MetaData'][0]['Codigo'])
            NumReliList.append(dato['Data'][i]['Valor'])       

ResidenciasMayores11 = pd.DataFrame({
        'Codigo Municipio' : CodResiList,
        'Nº Residencias Mayores': NumResiList
    })

InstitucionesReliMili11 = pd.DataFrame({
        'Codigo Municipio' : CodReliList,
        'Nº Instituciones Religiosas y Militares': NumReliList
    })

EstablecimientosCoelctivos11 = pd.concat([ResidenciasMayores11, InstitucionesReliMili11], axis = 1)
EstablecimientosCoelctivos11 = EstablecimientosCoelctivos11.\
                               loc[:, ~EstablecimientosCoelctivos11.columns.duplicated()]

EstablecimientosCoelctivos11.head()

EstablecimientosCoelctivos11.to_csv('13_Output_Establecimientos_Colectivos_11.csv', header = True, index = False)
