==> 2.Data/14_Code_Nacidos_Extranjero_11.ipynb <==
## Población nacida en el extranjero 2011

En este notebook procesaremos un JSON que contiene el número de personas; hombres y mujeres, nacidas en el extranjero desagredadas por municipios.

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº Hombres Extranjeros
- Nº Mujeres Extranjeros
- Nº Total Extranjeros

import pandas as pd
import requests

Path = 'https://servicios.ine.es/wstempus/js/es/DATOS_TABLA//t20/e244/avance/p02/l0/2mun00.px?tip=AM&p=2&'

json_request = requests.get(Path).json()

CodHomExtiList = list()
HomExtiList = list()
CodMujExtiList = list()
MujExtiList = list()
CodTotalExtiList = list()
TotalExtiList = list()

for dato in json_request: 
    for i in range(0,len(dato['Data'])):
        if 'Hombres, Nacidos en el extranjero' in dato['Nombre']:
            CodHomExtiList.append(dato['MetaData'][0]['Codigo'])
            HomExtiList.append(dato['Data'][i]['Valor'])
        elif 'Mujeres, Nacidos en el extranjero' in dato['Nombre']:
            CodMujExtiList.append(dato['MetaData'][0]['Codigo'])
            MujExtiList.append(dato['Data'][i]['Valor'])     
        elif 'Ambos sexos, Nacidos en el extranjero' in dato['Nombre']:
            CodTotalExtiList.append(dato['MetaData'][0]['Codigo'])
            TotalExtiList.append(dato['Data'][i]['Valor']) 

HombresExtranjeros11 = pd.DataFrame({
        'Codigo Municipio' : CodHomExtiList,
        'Nº Hombres Extranjeros': HomExtiList
    })

MujeresExtranjeras11 = pd.DataFrame({
        'Codigo Municipio' : CodMujExtiList,
        'Nº Mujeres Extranjeras': MujExtiList
    })

TotalExtranjeros11 = pd.DataFrame({
        'Codigo Municipio' : CodTotalExtiList,
        'Nº Total Extranjeros': TotalExtiList
    })

NacidosExtranjero11 = pd.concat([HombresExtranjeros11, MujeresExtranjeras11, TotalExtranjeros11], axis = 1)
NacidosExtranjero11 = NacidosExtranjero11.loc[:, ~NacidosExtranjero11.columns.duplicated()]

NacidosExtranjero11.head()

NacidosExtranjero11.to_csv('14_Output_Nacidos_Extranjero_11.csv', header = True, index = False)
