==> 2.Data/15_Code_Nacidos_Otra_CCAA_11.ipynb <==
## Población nacida en otra CCAA 2011

En este notebook procesaremos un JSON que contiene el número de personas; hombres y mujeres, nacidas en otra comunidad autónoma desagredadas por municipios.

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº Hombres Otra CCAA
- Nº Mujeres Otra CCAA
- Nº Total Otra CCAA

import pandas as pd
import requests

Path = 'https://servicios.ine.es/wstempus/js/es/DATOS_TABLA//t20/e244/avance/p02/l0/2mun00.px?tip=AM&p=2&'

json_request = requests.get(Path).json()

CodHomCCAAList = list()
HomCCAAList = list()
CodMujCCAAList = list()
MujCCAAList = list()
CodTotalCCAAList = list()
TotalCCAAList = list()

for dato in json_request: 
    for i in range(0,len(dato['Data'])):
        if 'Hombres, En distinta Comunidad Autónoma' in dato['Nombre']:
            CodHomCCAAList.append(dato['MetaData'][0]['Codigo'])
            HomCCAAList.append(dato['Data'][i]['Valor'])
        elif 'Mujeres, En distinta Comunidad Autónoma' in dato['Nombre']:
            CodMujCCAAList.append(dato['MetaData'][0]['Codigo'])
            MujCCAAList.append(dato['Data'][i]['Valor'])     
        elif 'Ambos sexos, En distinta Comunidad Autónoma' in dato['Nombre']:
            CodTotalCCAAList.append(dato['MetaData'][0]['Codigo'])
            TotalCCAAList.append(dato['Data'][i]['Valor']) 

HombresCCAA11 = pd.DataFrame({
        'Codigo Municipio' : CodHomCCAAList,
        'Nº Hombres Otra CCAA': HomCCAAList
    })

MujeresCCAA11 = pd.DataFrame({
        'Codigo Municipio' : CodMujCCAAList,
        'Nº Mujeres Otra CCAA': MujCCAAList
    })

TotalCCAA11 = pd.DataFrame({
        'Codigo Municipio' : CodTotalCCAAList,
        'Nº Total Otra CCAA': TotalCCAAList
    })

NacidosOtraCCAA11 = pd.concat([HombresCCAA11, MujeresCCAA11, TotalCCAA11], axis = 1)
NacidosOtraCCAA11 = NacidosOtraCCAA11.loc[:, ~NacidosOtraCCAA11.columns.duplicated()]

NacidosOtraCCAA11.head()

NacidosOtraCCAA11.to_csv('15_Output_Nacidos_Otra_CCAA_11.csv', header = True, index = False)
