==> 2.Data/16_Code_Edificios_Inmuebles_11.ipynb <==
## Número de edificios e inmuebles 2011

En este notebook procesaremos un JSON que contiene el número de edificos e inmuebles desagregado por municipios.

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº edificios
- Nº inmuebles

import pandas as pd
import requests

Path = 'https://servicios.ine.es/wstempus/js/es/DATOS_TABLA//t20/e244/edificios/p04/l0/1mun00.px?tip=AM&p=2&'

json_request = requests.get(Path).json()

CodEdiList = list()
EdiList = list()
CodInmuList = list()
InmuList = list()

for dato in json_request: 
    for i in range(0,len(dato['Data'])):
        if 'edificios' in dato['Nombre']:
            CodEdiList.append(dato['MetaData'][0]['Codigo'])
            EdiList.append(dato['Data'][i]['Valor'])
        elif 'inmuebles' in dato['Nombre']:
            CodInmuList.append(dato['MetaData'][0]['Codigo'])
            InmuList.append(dato['Data'][i]['Valor'])     

Edificios11 = pd.DataFrame({
        'Codigo Municipio' : CodEdiList,
        'Nº edificios': EdiList
    })

Inmuebles11 = pd.DataFrame({
        'Codigo Municipio' : CodInmuList,
        'Nº Inmuebles': InmuList
    })

EdificiosInmuebles11 = pd.concat([Edificios11, Inmuebles11], axis = 1)
EdificiosInmuebles11 = EdificiosInmuebles11.loc[:, ~EdificiosInmuebles11.columns.duplicated()]

EdificiosInmuebles11.head()

EdificiosInmuebles11.to_csv('16_Output_Edificios_Inmuebles_11.csv', header = True, index = False)
