==> 2.Data/17_Code_Viviendas_11.ipynb <==
## Número de viviendas 2011

En este notebook procesaremos un JSON que contiene el número de viviendas por municipio.

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº viviendas

import pandas as pd
import requests

Path = 'https://servicios.ine.es/wstempus/js/es/DATOS_TABLA//t20/e244/viviendas/p06/l0/9mun00.px?tip=AM&p=2&'

json_request = requests.get(Path).json()

CodViviList = list()
ViviList = list()

for dato in json_request: 
    for i in range(0,len(dato['Data'])):
        if 'viviendas familiares' in dato['Nombre']:
            CodViviList.append(dato['MetaData'][0]['Codigo'])
            ViviList.append(dato['Data'][i]['Valor'])    

Viviendas11 = pd.DataFrame({
        'Codigo Municipio' : CodViviList,
        'Nº viviendas': ViviList
    })

Viviendas11.head()

Viviendas11.to_csv('17_Output_Viviendas_11.csv', header = True, index = False)
