==> 2.Data/18_Code_Edad_11.ipynb <==
## Edades 2011
En este notebook procesaremos un .xlsx con el número de personas por tramos de edad desagregado por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Edad < 16 	
- 16 =< Edad =<64	
- Edad >= 65

import pandas as pd

Edades11 = pd.read_excel('18_Input_Edad_11.xlsx', dtype = {'Municipio de residencia' : str})

Edades11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str,
                                                        'Codigo Provincia' : str,
                                                         'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

Edades11['Codigo Municipio'] = Edades11['Municipio de residencia'].map(lambda x: check_list(x))

Edades11.head()

Edades11.rename({'Menos de 16' : 'Edad < 16', 
                        '16-64' : '16 =< Edad =<64',
                       '65 ó más' : 'Edad >= 65'}, axis = 1, inplace = True)
ColNue = ['Codigo Municipio', 'Edad < 16', '16 =< Edad =<64', 'Edad >= 65']
Edades11 = Edades11[ColNue]

Edades11.head()

Edades11.to_csv('18_Output_Edad_11.csv', header = True, index = False)
