==> 2.Data/19_Code_Estado_Civil_11.ipynb <==
## Estado civil 2011
En este notebook procesaremos un .xlsx con el número de personas segun su situacion civil desagregado por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Soltero/a 	
- Casado/a
- Separado/a
- Divorciado/a
- Viudo/a

import pandas as pd

EstadoCivil11 = pd.read_excel('19_Input_Estado_Civil_11.xlsx', dtype = {'Municipio de residencia' : str})

EstadoCivil11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str,
                                                         'Codigo Provincia' : str, 
                                                         'Codigo Municipio' : str }) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

EstadoCivil11['Codigo Municipio'] = EstadoCivil11['Municipio de residencia'].map(lambda x: check_list(x))

EstadoCivil11.head()

ColNue = ['Codigo Municipio', 'Soltero/a', 'Casado/a', 'Separado/a', 'Divorciado/a', 'Viudo/a']
EstadoCivil11 = EstadoCivil11[ColNue]

EstadoCivil11.head()

EstadoCivil11.to_csv('19_Output_Estado_Civil_11.csv', header = True, index = False)
