==> 2.Data/20_Code_Nivel_Estudios_11.ipynb <==
## Nivel de estudios 2011
En este notebook procesaremos un .xlsx con el número de personas segun su nivel de estudios desagregado por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Sin Estudios 
- Con Estudios

import pandas as pd

NivelEstudios11 = pd.read_excel('20_Input_Nivel_Estudios_11.xlsx', dtype = {'Municipio de residencia' : str})

NivelEstudios11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str,
                                                        'Codigo Provincia' : str, 
                                                         'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

NivelEstudios11['Codigo Municipio'] = NivelEstudios11['Municipio de residencia'].map(lambda x: check_list(x))

NivelEstudios11.head()

Juntamos las columnas Estudios Primarios o Secundarios y Estudios Superiores en una única columna.  
Como los valores que están entre 0 y 5 se indican en la tabla con un asterisco, sustituimos estos valores por un 2.

ColumnasAsterisco = NivelEstudios11.columns[NivelEstudios11.isin(['*']).any()]

ColumnasAsterisco

NivelEstudios11[ColumnasAsterisco] = NivelEstudios11[ColumnasAsterisco].replace('*', 2)

NivelEstudios11['Con Estudios'] = NivelEstudios11['Estudios Primarios o Secundarios'] +\
                                  NivelEstudios11['Estudios Superiones']

NivelEstudios11.head()

NivelEstudios11 = NivelEstudios11.drop(['Municipio de residencia',
                                       'Estudios Primarios o Secundarios',
                                       'Estudios Superiones'],
                                        axis = 1)
PrimeraColumna = NivelEstudios11.pop('Codigo Municipio')
NivelEstudios11.insert(0,'Codigo Municipio', PrimeraColumna)

NivelEstudios11.head()

NivelEstudios11.to_csv('20_Output_Nivel_Estudios_11.csv', header = True, index = False)
