==> 2.Data/21_Code_Jubilados_11.ipynb <==
## Número de jubilados 2011
En este notebook procesaremos un .xlsx con el número de personas jubiladas desagregado por municipios.  
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº de Jubilados 	

import pandas as pd

Jubilados11 = pd.read_excel('21_Input_Jubilados_11.xlsx', 
                                       usecols = ['Municipio de residencia',
                                                'Jubilado/a, prejubilado/a, pensionista o rentista'], 
                                       dtype = {'Municipio de residencia' : str})

Jubilados11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str,
                                                         'Codigo Provincia' : str,
                                                         'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

Jubilados11['Codigo Municipio'] = Jubilados11['Municipio de residencia'].map(lambda x: check_list(x))

Jubilados11.head()

Jubilados11 = Jubilados11.drop(['Municipio de residencia'], axis = 1)
Jubilados11 = Jubilados11.rename({'Jubilado/a, prejubilado/a, pensionista o rentista': 'Nº de Jubilados'}, 
                                 axis = 1)
PrimeraColumna = Jubilados11.pop('Codigo Municipio')
Jubilados11.insert(0,'Codigo Municipio', PrimeraColumna)

Jubilados11.head()

Jubilados11.to_csv('21_Output_Jubliados_11.csv', header = True, index = False)
