==> 2.Data/22_Code_Empresarios_11.ipynb <==
## Número de empresarios 2011
En este notebook procesaremos un .xlsx con el número de empresarios, tanto el que tiene empleados como el que no tiene, desagregado por municipios. 

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Empresario que emplea personal 	
- Empresario que no emplea personal

import pandas as pd

Empresarios11 = pd.read_excel('22_Input_Empresarios_11.xlsx', 
                              usecols = ['Municipio de residencia',
                                        'Empresario que emplea personal',
                                         'Empresario que no emplea personal'], 
                              dtype = {'Municipio de residencia' : str})

Empresarios11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str, 
                                                         'Codigo Provincia' : str,
                                                         'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

Empresarios11['Codigo Municipio'] = Empresarios11['Municipio de residencia'].map(lambda x: check_list(x))

Empresarios11.head()

Empresarios11 = Empresarios11.drop(['Municipio de residencia'], axis = 1)
PrimeraColumna = Empresarios11.pop('Codigo Municipio')
Empresarios11.insert(0,'Codigo Municipio', PrimeraColumna)

Empresarios11.head()

Empresarios11.to_csv('22_Output_Empresarios_11.csv', header = True, index = False)
