==> 2.Data/23_Code_Numero_Hijos_11.ipynb <==
## Número de hijos 2011
En este notebook procesaremos un .xlsx con el número de hijos desagregado por municipios. 

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- 1 hijo/a 
- 2 hijos/as 	
- 3 hijos/as 	
- 4 o más hijos/as

import pandas as pd

Hijos11 = pd.read_excel('23_Input_Numero_Hijos_11.xlsx', dtype = {'Municipio de residencia' : str})

Hijos11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str, 
                                                         'Codigo Provincia' : str,
                                                         'Codigo Municipio' :str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

Hijos11['Codigo Municipio'] = Hijos11['Municipio de residencia'].map(lambda x: check_list(x))

Hijos11.head()

Hijos11 = Hijos11.drop(['Municipio de residencia'], axis = 1)
PrimeraColumna = Hijos11.pop('Codigo Municipio')
Hijos11.insert(0,'Codigo Municipio', PrimeraColumna)

Hijos11.head()

Hijos11.to_csv('23_Output_Numero_Hijos_11.csv', header = True, index = False)
