==> 2.Data/24_Code_Alquileres_11.ipynb <==
## Número de alquileres 2011
En este notebook procesaremos un .xlsx con el número de viviendas en régime de alquiler desagregado por municipios. 

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº de Alquileres

import pandas as pd

Alquileres11 = pd.read_excel('24_Input_Alquileres_11.xlsx', usecols = ['Municipio de residencia',
                                                                      'Alquilada'], 
                             dtype = {'Municipio de residencia' : str})

Alquileres11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str,
                                                         'Codigo Provincia' : str,
                                                         'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

Alquileres11['Codigo Municipio'] = Alquileres11['Municipio de residencia'].map(lambda x: check_list(x))

Alquileres11.head()

Alquileres11 = Alquileres11.drop(['Municipio de residencia'], axis = 1)
Alquileres11 = Alquileres11.rename({'Alquilada' : 'Nº de Alquileres'}, axis = 1)
PrimeraColumna = Alquileres11.pop('Codigo Municipio')
Alquileres11.insert(0,'Codigo Municipio', PrimeraColumna)

Alquileres11.head()

Alquileres11.to_csv('24_Output_Alquileres_11.csv', header = True, index = False)
