==> 2.Data/25_Code_Tipos_Hogar_11.ipynb <==
## Tipos de hogar 2011
En este notebook procesaremos un .xlsx con el número de hogares unipersonales y familiares desagregado por municipios. 

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Hogares Unipersonales
- Hogares Familiares

import pandas as pd

TiposHogar11 = pd.read_excel('25_Input_Tipos_Hogar_11.xlsx', dtype = {'Municipios de residencia' : str})

TiposHogar11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype={'Codigo CCAA' : str,
                                                         'Codigo Provincia' : str,
                                                         'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

TiposHogar11['Codigo Municipio'] = TiposHogar11['Municipio de residencia'].map(lambda x: check_list(x))

TiposHogar11.head()

TiposHogar11 = TiposHogar11.drop(['Municipio de residencia'], axis = 1)
PrimeraColumna = TiposHogar11.pop('Codigo Municipio')
TiposHogar11.insert(0,'Codigo Municipio', PrimeraColumna)

TiposHogar11.head()

TiposHogar11.to_csv('25_Output_Tipos_Hogar_11.csv', header = True, index = False)
