==> 2.Data/26_Code_Personas_Hogar_11.ipynb <==
## Número de personas en el hogar 2011
En este notebook procesaremos un .xlsx con el número de personas por hogar desagregado por municipios. 

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- 1 persona 	
- 2 personas 	
- 3 personas 	
- 4 personas 	
- 5 personas 	
- 6 o más personas

import pandas as pd

PersonasHogar11 = pd.read_excel('26_Input_Personas_Hogar_11.xlsx', dtype = {'Municipio de residencia' : str})

PersonasHogar11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str,
                                                           'Codigo Provincia' : str,
                                                           'Codigo Municipio' : str}) 
ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

PersonasHogar11['Codigo Municipio'] = PersonasHogar11['Municipio de residencia'].map(lambda x: check_list(x))

PersonasHogar11.head()

PersonasHogar11 = PersonasHogar11.drop(['Municipio de residencia'], axis = 1)
PrimeraColumna = PersonasHogar11.pop('Codigo Municipio')
PersonasHogar11.insert(0,'Codigo Municipio', PrimeraColumna)

PersonasHogar11.head()

PersonasHogar11.to_csv('26_Output_Personas_Hogar_11.csv', header = True, index = False)
