==> 2.Data/27_Code_Edad_Edificios_11.ipynb <==
## Edad medida de los edificios 2011
En este notebook procesaremos un .xlsx con la edad media de los edificios desagregado por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Edad Media Edificios

import pandas as pd
import numpy as np

EdadEdificios11 = pd.read_excel('27_Input_Edad_Edificios_11.xlsx', dtype = {'Código' : str})

EdadEdificios11.head()

EdadEdificios11.drop(['Etiqueta'], axis = 1, inplace = True)
EdadEdificios11.rename({'Código' : 'Codigo Municipio',
                        'Edad media de los edificios 2011' : 'Edad Media Edificios'}, axis = 1, inplace = True)

EdadEdificios11.head()

Algunas variables toman el valor: N/A - resultado no disponible.  
Lo sustituimos por NaN:

EdadEdificios11.replace('N/A - resultado no disponible', np.nan, inplace = True)

EdadEdificios11.to_csv('27_Output_Edad_Edificios_11.csv', header = True, index = False)
