==> 2.Data/28_Code_Uso_suelo_11.ipynb <==
## Uso del suelo 2011

En este notebook procesaremos un .xlsx con el número de hectáreas de uso del suelo; industrial, residencial, terciario y otros, desagregado por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Uso Industrial
- Uso Residelcial
- Uso Terciario
- Otros Usos


import pandas as pd
import numpy as np

UsoSuelo11 = pd.read_excel('28_Input_Uso_Suelo_11.xlsx', dtype = {'Código' : str})

UsoSuelo11.head()

UsoSuelo11.drop('Etiqueta', axis = 1, inplace = True)
UsoSuelo11.rename({'Código':'Codigo Municipio',
                  'industrial' : 'Uso Industrial',
                  'residencial' : 'Uso Residelcial',
                  'terciario' : 'Uso Terciario',
                  'otros' : 'Otros Usos'}, axis = 1, inplace = True)

UsoSuelo11.head()

Algunas variables toman el valor: N/A - resultado no disponible.  
Lo sustituimos por NaN:

UsoSuelo11.replace('N/A - resultado no disponible', np.nan, inplace = True)

UsoSuelo11.to_csv('28_Output_Uso_Suelo_11.csv', header = True, index = False)
