==> 2.Data/29_Code_Acceso_Internet_11.ipynb <==
## Número de hogares que tienen o no internet en 2011
En este notebook procesaremos un .xlsx con el número de hogares que tienen o no internet desagregado por municipios. 

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº hogares que tiene internet
- Nº hogares que no tienen internet

import pandas as pd

AccesoInternet11 = pd.read_excel('29_Input_Acceso_Internet_11.xlsx', dtype = {'Municipio de residencia' : str})

AccesoInternet11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.  
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio:

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str,
                                                         'Codigo Provincia' : str,
                                                         'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

AccesoInternet11['Codigo Municipio'] = AccesoInternet11['Municipio de residencia'].map(lambda x: check_list(x))

AccesoInternet11.head()

AccesoInternet11 = AccesoInternet11.drop(['Municipio de residencia'], axis = 1)
PrimeraColumna = AccesoInternet11.pop('Codigo Municipio')
AccesoInternet11.insert(0,'Codigo Municipio', PrimeraColumna)

AccesoInternet11.head()

AccesoInternet11.to_csv('29_Output_Acceso_Internet_11.csv', header = True, index = False)
