==> 2.Data/30_Code_Parque_Vehiculos_11.ipynb <==
## Parque de vehículos 2011

En este notebook procesaremos un .xlsx con el número de vehículos por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº vehículos



import pandas as pd

Vehiculos11 = pd.read_excel('30_Input_Parque_Vehiculos_11.xlsx', dtype = {'Codigo Municipio' : str})

Vehiculos11.head()

Vehiculos11.drop('Municipio de residencia', axis = 1, inplace = True)
Vehiculos11.rename({'Total' : 'Nº Vehiculos'}, axis = 1, inplace = True)

Vehiculos11.head()

Vehiculos11.to_csv('30_Output_Parque_Vehiculos_11.csv', header = True, index = False)
