==> 2.Data/31_Code_Censo_Conductores_11.ipynb <==
## Censo conductores 2011

En este notebook procesaremos un .xlsx con el número de conductores por municipios.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº de conductores



import pandas as pd

Conductores11 = pd.read_excel('31_Input_Censo_Conductores_11.xlsx', dtype = {'Codigo Municipio' : str})

Conductores11.head()

Conductores11.drop('Municipio de residencia', axis = 1, inplace = True)
Conductores11.rename({'Total' : 'Nº Conductores'}, axis = 1, inplace = True)

Conductores11.head()

Conductores11.to_csv('31_Output_Censo_Conductores_11.csv', header = True, index = False)
