==> 2.Data/32_Code_Deuda_Ayuntamientos_11.ipynb <==
## Deuda ayuntamientos 2011

En este notebook procesaremos un .xlsx con la deuda viva de cada uno de los municipios en 2011.  

Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Deuda (Miles de €)



import pandas as pd

Deuda11 = pd.read_excel('32_Input_Deuda_Ayuntamientos_11.xlsx', dtype = {'Codigo Municipio' : str})

Deuda11.head()

Deuda11.drop(['Codigo CA', 
              'Codigo Provin',
              'Municipio'], axis = 1, inplace = True)

Deuda11.head()

Deuda11.to_csv('32_Output_Deuda_Ayuntamientos_11.csv', header = True, index = False)
