==> 2.Data/33_Code_Impuestos_11.ipynb <==
## Impuestos ayuntamientos 2011
En este notebook procesaremos un .xlsx con el total de impuestos que recauda cada municipio.  
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Total impuestos (Miles €)

import pandas as pd

Impuestos11 = pd.read_excel('33_Input_Impuestos_11.xlsx', dtype = {'Prov' : str,'Corp' : str}, 
                            usecols = ['Prov', 'Corp', 'Impuestos directos', 'Impuestos indirectos'])

Impuestos11.head()

Creamos el campo Codigo Municipio

Impuestos11['Codigo Municipio'] = Impuestos11['Prov'].map(str)+Impuestos11['Corp'].map(str)

Impuestos11.head()

Creamos una única columna con la suma de los impuestos directos e indirectos y lo pasamos a Miles de euros

Impuestos11['Total impuestos (Miles de €)'] = (Impuestos11['Impuestos directos'] \
                                                   + Impuestos11['Impuestos indirectos'])/1000

Impuestos11.head()

Impuestos11.drop(['Prov','Corp', 'Impuestos directos','Impuestos indirectos'], axis = 1, inplace = True)
PrimeraColumna = Impuestos11.pop('Codigo Municipio')
Impuestos11.insert(0,'Codigo Municipio', PrimeraColumna)

Impuestos11.head()

Impuestos11.to_csv('33_Output_Impuestos_11.csv', header = True, index = False)
