==> 2.Data/34_Code_Explotaciones_Agricolas_11.ipynb <==
## Explotaciones agrícolas 2011
En este notebook procesaremos un .xlsx con el número de explotaciones y hectáreas agrícolas segmentadas por municipios.  
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Superficie (Ha) Agricolas
- Nº Explotaciones Agricolas

import pandas as pd

ExplotacionesAgricolas11 = pd.read_excel('34_Input_Explotaciones_Agricolas_11.xlsx',sheet_name = 'Explotaciones')

ExplotacionesAgricolas11.head()

SuperficieAgricolas11 = pd.read_excel('34_Input_Explotaciones_Agricolas_11.xlsx',sheet_name = 'Superficie')

SuperficieAgricolas11.head()

Unimos ambas tablas a través del campo Municipio:

Agricultura11 = ExplotacionesAgricolas11.merge(SuperficieAgricolas11, on = 'Municipio')

Agricultura11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str,
                                                           'Codigo Provincia' : str,
                                                           'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

Agricultura11['Codigo Municipio'] = Agricultura11['Municipio'].map(lambda x: check_list(x))

Agricultura11.head()

PrimeraColumna = Agricultura11.pop('Codigo Municipio')
Agricultura11.insert(0,'Codigo Municipio', PrimeraColumna)

Agricultura11.head()

Calculamos el total de explotaciones:

Agricultura11['Nº Explotaciones Agrícolas'] = Agricultura11['SAU'] \
                                                + Agricultura11['Otras tierras']

Agricultura11.drop(['Municipio', 'SAU', 'Otras tierras'], axis = 1, inplace = True)
Agricultura11.rename({'Superficie (Ha)' : 'Superficie (Ha) Agricolas'}, axis = 1, inplace = True)

Agricultura11.head()

Agricultura11.to_csv('34_Output_Explotaciones_Agricolas_11.csv', header = True, index = False)
