==> 2.Data/35_Code_Explotaciones_Ganaderas_11.ipynb <==
## Explotaciones ganaderas 2011
En este notebook procesaremos un .xlsx con el número de explotaciones ganaderas segmentadas por municipios.  
Obtendremos un DataFrame con la siguiente información:
- Codigo Municipio
- Nº Explotaciones Ganaderas

import pandas as pd
import numpy as np

ExplotacionesGanaderas11 = pd.read_excel('35_Input_Explotaciones_Ganaderas_11.xlsx')

ExplotacionesGanaderas11.head()

Añadimos el campo Código Municipio al DataFrame asignando el valor de cada fila en función del código que aparece en el campo municipio de residencia.
Para ello usamos la tabla 01_Output_ProvMun_11.csv que contiene los códigos por municipio

ProvMun = pd.read_csv('01_Output_ProvMun_11.csv', dtype = {'Codigo CCAA' : str, 
                                                           'Codigo Provincia' : str,
                                                           'Codigo Municipio' : str}) 

ProvMun.head()

MunList = ProvMun['Codigo Municipio'].tolist()

def check_list(x):
    for l in MunList:
        if l in x:
            return l
    return ''

ExplotacionesGanaderas11['Codigo Municipio'] = ExplotacionesGanaderas11['Municipio'].map(lambda x: check_list(x))

ExplotacionesGanaderas11.head()

PrimeraColumna = ExplotacionesGanaderas11.pop('Codigo Municipio')
ExplotacionesGanaderas11.insert(0,'Codigo Municipio', PrimeraColumna)

ExplotacionesGanaderas11.head()

Calculamos el total de explotaciones ganaderas.  
Para ello, sustituimos los NaN por ceros.

ExplotacionesGanaderas11 = ExplotacionesGanaderas11.replace(np.nan, 0)

ExplotacionesGanaderas11['Nº Explotaciones Ganaderas'] = ExplotacionesGanaderas11['Bovinos'] \
                                                        + ExplotacionesGanaderas11['Ovinos'] \
                                                        + ExplotacionesGanaderas11['Caprinos'] \
                                                        + ExplotacionesGanaderas11['Equinos (caballos, mulas y asnos)'] \
                                                        + ExplotacionesGanaderas11['Porcinos'] \
                                                        + ExplotacionesGanaderas11['Aves'] \
                                                        + ExplotacionesGanaderas11['Conejas madres (sólo hembras reproductoras)']\
                                                        + ExplotacionesGanaderas11['Colmenas']
                                                        
ExplotacionesGanaderas11.drop(['Municipio', 'Bovinos', 'Ovinos', 'Caprinos',
                              'Equinos (caballos, mulas y asnos)', 'Porcinos',
                              'Aves', 'Conejas madres (sólo hembras reproductoras)', 'Colmenas'],\
                              axis = 1, inplace = True)

ExplotacionesGanaderas11.head()

ExplotacionesGanaderas11.to_csv('35_Output_Explotaciones_Ganaderas_11.csv', header = True, index = False)
